pyproject.toml
[project]
name = "testovoe"
version = "0.1.0"
description = "Mini auth API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",
    "bcrypt>=4.1.0",
    "PyJWT>=2.8.0",
    "structlog>=24.1.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.19.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["venv", ".venv"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
exclude = ["^venv/", "^\\.venv/"]
